[integration-test] Build image on push/PR, push to ghcr.io on release
Adjust the workflow to rebuild the integration-test image whenever a new
commit is pushed to the repository a pull request is opened that touches
any path in the `integration-test/` subdirectory. If the workflow is
triggered by a release being published, the built image is additionally
pushed to ghcr.io.

This now allows PRs and work branches to check if the image builds
cleanly before cutting a new release.